feat: sign in using Microsoft account
I created the Swift package to wrap the Microsoft Authentication Library
(MSAL) and provide a friendlier Swift/SwiftUI interface. I created the
AuthenticationError type to better represent errors and provide access
to the error details.

I created the MicrosoftAccountStore class to be used by a SwiftUI
application to log in using a Microsoft account and obtain an access
token to call APIs with.